// apps/frontend/src/types/payload-types.ts
export type SkiWaiverFormDependent =
  | {
      name: string;
      qid?: string | null;
      age: number;
      medical_conditions?: string | null;
      id?: string | null;
    }[]
  | null;
// ...
export type SkiWaiverContentDetail =
  | {
      title?: string | null;
      description?:
        | {
            [k: string]: unknown;
          }[]
        | null;
      id?: string | null;
    }[]
  | null;
// ...
export interface User {
  id: number;
  operator?: (number | null) | Operator;
  role?: (number | null) | UserRole;
  kpi_department?: (number | null) | KpiPlansDepartment;
  restaurant?: (number | null) | Restaurant;
  access?: (number | null) | UserAccess;
  auth_method?: ("default" | "microsoft") | null;
  updatedAt: string;
  createdAt: string;
  enableAPIKey?: boolean | null;
  apiKey?: string | null;
  apiKeyIndex?: string | null;
  email: string;
  resetPasswordToken?: string | null;
  resetPasswordExpiration?: string | null;
  salt?: string | null;
  hash?: string | null;
  loginAttempts?: number | null;
  lockUntil?: string | null;
  password: string | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "operators".
 */
export interface Operator {
  id: number;
  name?: string | null;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "user-roles".
 */
export interface UserRole {
  id: number;
  name?: string | null;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "kpi-plans-department".
 */
export interface KpiPlansDepartment {
  id: number;
  created_by?: (number | null) | User;
  updated_by?: (number | null) | User;
  name: string;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "restaurants".
 */
export interface Restaurant {
  id: number;
  logo?: number | Media | null;
  featured_image?: number | Media | null;
  name: string;
  slug: string;
  description?: string | null;
  description_rt?:
    | {
        [k: string]: unknown;
      }[]
    | null;
  options?: {
    include_survey_customer?: boolean | null;
    include_employee_perks?: boolean | null;
    order?: number | null;
  };
  scripts?: HeadScripts;
  additional_data?:
    | {
        [k: string]: unknown;
      }
    | unknown[]
    | string
    | number
    | boolean
    | null;
  created_by?: (number | null) | User;
  updated_by?: (number | null) | User;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "media".
 */
export interface Media {
  id: number;
  alt?: string | null;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "user-access".
 */
export interface UserAccess {
  id: number;
  name?: string | null;
  access?:
    | {
        name?: string | null;
        hidden?: boolean | null;
        read?: boolean | null;
        create?: boolean | null;
        update?: boolean | null;
        admin?: boolean | null;
        delete?: boolean | null;
        access?: boolean | null;
        id?: string | null;
      }[]
    | null;
  updatedAt: string;
  createdAt: string;
}
// ...
export interface SkiWaiverForm {
  id: number;
  date?: string | null;
  name: string;
  qid: string;
  phone: string;
  age: number;
  emergency_contact_name: string;
  emergency_contact_phone: string;
  dependents?: SkiWaiverFormDependent;
  signature_svg: string;
  updatedAt: string;
  createdAt: string;
}
// ...
export interface SkiWaiverContent {
  id: number;
  primary_logo?: number | Media | null;
  secondary_logo?: number | Media | null;
  title?: string | null;
  details?: SkiWaiverContentDetail;
  created_by?: (number | null) | User;
  updated_by?: (number | null) | User;
  _status?: ("draft" | "published") | null;
  updatedAt?: string | null;
  createdAt?: string | null;
}